test(WinningNumbersTest): move tests for has() method of LottoGame to WinningNumbersTest and remove has() from LottoGame
